requirements.txt
torch
transformers
datasets
google-genai
python-dotenv
bitsandbytes
peft
trl
